Move plugin ABIs to shared directory
