+++ src/plugins/abi/ANYSWAP_V5_ERC20_ABI.json
