+++ src/plugins/abi/STABLE_DEBT_TOKEN_ABI.json
