+++ src/plugins/abi/UNISWAP_V2_PAIR.json
