+++ src/plugins/abi/VELODROME_V2_GAUGE.json
