+++ src/plugins/stake-plugins/uniswapV2/policyInfo/fantom.ts
@@ -1,12 +1,12 @@
 import { ENV } from '../../../../env'
+import ANYSWAP_V5_ERC20_ABI from '../../../abi/ANYSWAP_V5_ERC20_ABI.json'
+import MASONRY_ABI from '../../../abi/MASONRY_ABI.json'
+import TOMB_TREASURY_ABI from '../../../abi/TOMB_TREASURY_ABI.json'
+import TSHARE_REWARD_POOL_ABI from '../../../abi/TSHARE_REWARD_POOL_ABI.json'
+import UNISWAP_V2_PAIR from '../../../abi/UNISWAP_V2_PAIR.json'
+import UNISWAP_V2_ROUTER_02 from '../../../abi/UNISWAP_V2_ROUTER_02.json'
+import WRAPPED_FTM_ABI from '../../../abi/WRAPPED_FTM_ABI.json'
 import { StakeProviderInfo } from '../../types'
-import ANYSWAP_V5_ERC20_ABI from '../abi/ANYSWAP_V5_ERC20_ABI.json'
-import MASONRY_ABI from '../abi/MASONRY_ABI.json'
-import TOMB_TREASURY_ABI from '../abi/TOMB_TREASURY_ABI.json'
-import TSHARE_REWARD_POOL_ABI from '../abi/TSHARE_REWARD_POOL_ABI.json'
-import UNISWAP_V2_PAIR from '../abi/UNISWAP_V2_PAIR.json'
-import UNISWAP_V2_ROUTER_02 from '../abi/UNISWAP_V2_ROUTER_02.json'
-import WRAPPED_FTM_ABI from '../abi/WRAPPED_FTM_ABI.json'
 import { makeEcosystem } from '../Ecosystem'
 import { makeCemeteryPolicy } from '../policies/cemeteryPolicy'
 import { makeMasonryPolicy } from '../policies/masonryPolicy'
