+++ src/plugins/stake-plugins/uniswapV2/policyInfo/optimism.ts
@@ -1,9 +1,9 @@
+import ANYSWAP_V5_ERC20_ABI from '../../../abi/ANYSWAP_V5_ERC20_ABI.json'
+import UNISWAP_V2_PAIR from '../../../abi/UNISWAP_V2_PAIR.json'
+import VELODROME_V2_GAUGE from '../../../abi/VELODROME_V2_GAUGE.json'
+import VELODROME_V2_ROUTER from '../../../abi/VELODROME_V2_ROUTER.json'
+import WRAPPED_ETH_ABI from '../../../abi/WRAPPED_ETH_ABI.json'
 import { StakeProviderInfo } from '../../types'
-import ANYSWAP_V5_ERC20_ABI from '../abi/ANYSWAP_V5_ERC20_ABI.json'
-import UNISWAP_V2_PAIR from '../abi/UNISWAP_V2_PAIR.json'
-import VELODROME_V2_GAUGE from '../abi/VELODROME_V2_GAUGE.json'
-import VELODROME_V2_ROUTER from '../abi/VELODROME_V2_ROUTER.json'
-import WRAPPED_ETH_ABI from '../abi/WRAPPED_ETH_ABI.json'
 import { makeEcosystem } from '../Ecosystem'
 import { makeVelodromeV2StakePolicy } from '../policies/VelodromeV2StakePolicy'
 import { StakePolicyInfo } from '../stakePolicy'
